// fastlearnerfrontend/src/app/core/models/button.model-config.ts
export class buttonConfig {
  type?: any;
  size?: any;
  iconType?: string;
  backgroundColor?: string;
  color?: string;
  text?: string;
  borderColor?: string;
  outline?: string;
  border?: string;
  borderRadius?: string;
  height?: string;
  paddingTop?: string;
  paddingRight?: string;
  paddingBottom?: string;
  paddingLeft?: string;
  fontSize?: any;
  hoverConfig?: HoverConfig;
  fontWeight?: string;
}

export class HoverConfig {
  color?: string;
  backgroundColor?: string
}